[SofaSimulationCore] Move MechanicalVInitVisitor to dedicated folder
